+++ algomancy/scenarioengine/scenariofactory.py
@@ -1,8 +1,8 @@
-from typing import Dict, List, Optional, Any
+from typing import Dict, List, Optional, Any, Type
 
 from algomancy.scenarioengine.algorithmfactory import AlgorithmFactory
 from algomancy.dashboardlogger.logger import Logger
-from algomancy.scenarioengine.algorithmtemplate import AlgorithmTemplate
+from algomancy.scenarioengine.basealgorithm import ALGORITHM
 from algomancy.scenarioengine.keyperformanceindicator import KpiTemplate, build_kpis
 from algomancy.scenarioengine.scenario import Scenario
 from algomancy.dataengine.datamanager import DataManager
@@ -13,11 +13,11 @@ class ScenarioFactory:
     Creates scenarios, builds algorithms and KPIs, and performs parameter validation.
     """
 
-    def __init__(self, kpi_templates: List[KpiTemplate], algo_templates: Dict[str, AlgorithmTemplate], data_manager: DataManager, logger: Logger | None = None):
+    def __init__(self, kpi_templates: List[KpiTemplate], algo_templates: Dict[str, Type[ALGORITHM]],
+                 data_manager: DataManager, logger: Logger | None = None):
         self.logger = logger
-        self._kpi_templates = kpi_templates
-        # self._algo_templates = algo_templates
-        self.algorithm_factory = AlgorithmFactory(algo_templates=algo_templates, logger=logger)
+        self._kpi_templates = kpi_templates  # todo this should become a KPIFactory
+        self._algorithm_factory = AlgorithmFactory(algo_templates, logger)
         self._data_manager = data_manager
 
     def log(self, msg: str):
@@ -26,7 +26,11 @@ class ScenarioFactory:
 
     @property
     def available_algorithms(self) -> List[str]:
-        return self.algorithm_factory.available_algorithms
+        return self._algorithm_factory.available_algorithms
+
+    @property
+    def algo_templates(self) -> Dict[str, Type[ALGORITHM]]:
+        return self._algorithm_factory.templates
 
     def create(self, tag: str, dataset_key: str, algo_name: str, algo_params: Optional[dict] = None) -> Scenario:
         if algo_params is None:
@@ -35,7 +39,7 @@ class ScenarioFactory:
         assert algo_name in self.available_algorithms, f"Algorithm '{algo_name}' not found."
         assert dataset_key in self._data_manager.get_data_keys(), f"Data '{dataset_key}' not found."
 
-        algorithm = self.algorithm_factory.create(
+        algorithm = self._algorithm_factory.create(
             input_name=algo_name,
             input_params=algo_params,
         )